// src/types/wizard/store.ts
import { Action, Thunk } from "easy-peasy";
import { WizardState, WizardStep } from "~/types/wizard/wizard";
import { QuestionAndResponse } from "~/types/api/questionAndResponse";
import { InterviewResponse } from "~/types/api/interviewResponse";
interface expandAll {
  expandAll: boolean;
}
export interface IWizardStoreModel {
  expandAll: expandAll;
  wizard: WizardState;
  activeStep: WizardStep;
  activeStepIndex: number;
  activeSubStep: WizardStep;
  activeSubStepIndex: number;
  activeQuestion: QuestionAndResponse;
  activeQuestionIndex: number;
  activeQuestionId: number;
  activeParentQuestionIndex: number;
  activeParentQuestionId: number;
  activeResponse: InterviewResponse;
  activeResponseIndex: number;
  activeResponseId: number;
  showStarredQuestionsOnly: boolean;
  showStarredResponsesOnly: boolean;
  toggleStarredQuestions: Action<IWizardStoreModel, unknown>;
  onToggleStarredQuestions: Thunk<IWizardStoreModel, unknown>;
  toggleStarredQuestion: Action<IWizardStoreModel, unknown>;
  onToggleStarredQuestion: Thunk<IWizardStoreModel, unknown>;
  onToggleStarredResponse: Thunk<IWizardStoreModel, unknown>;
  toggleStarredResponse: Action<IWizardStoreModel, unknown>;
  onToggleStarredResponses: Thunk<IWizardStoreModel, unknown>;
  toggleStarredResponses: Action<IWizardStoreModel, unknown>;
  onToggleHiddenResponse: Thunk<IWizardStoreModel, unknown>;
  toggleHiddenResponse: Action<IWizardStoreModel, unknown>;
  onToggleFilter: Thunk<IWizardStoreModel, unknown>;
  toggleFilter: Action<IWizardStoreModel, unknown>;
  showResponseForm: boolean;
  toggleShowResponseForm: Action<IWizardStoreModel, unknown>;
  onToggleShowResponseForm: Thunk<IWizardStoreModel, unknown>;
  showSubQuestions: boolean;
  toggleShowSubQuestions: Action<IWizardStoreModel, unknown>;
  toggleExpanded: Action<IWizardStoreModel, unknown>;
  setSingleExpanded: Action<IWizardStoreModel, unknown>;
  onToggleExpanded: Thunk<IWizardStoreModel, unknown>;
  onSetSingleExpanded: Thunk<IWizardStoreModel, unknown>;
  onToggleShowSubQuestions: Thunk<IWizardStoreModel, unknown>;
  populate: Action<IWizardStoreModel, unknown>;
  onPopulate: Thunk<IWizardStoreModel, unknown>;
  select: Action<IWizardStoreModel, unknown>;
  onSelect: Thunk<IWizardStoreModel, unknown>;
  onClear: Thunk<IWizardStoreModel, unknown>;
  clear: Action<IWizardStoreModel, unknown>;
  addResponse: Action<IWizardStoreModel, unknown>;
  onAddResponse: Thunk<IWizardStoreModel, unknown>;
  updateResponse: Action<IWizardStoreModel, unknown>;
  onUpdateResponse: Thunk<IWizardStoreModel, unknown>;
  removeResponse: Action<IWizardStoreModel, unknown>;
  onRemoveResponse: Thunk<IWizardStoreModel, unknown>;
  onSelectQuestion: Thunk<IWizardStoreModel, unknown>;
  selectQuestion: Action<IWizardStoreModel, unknown>;
  onReset: Thunk<IWizardStoreModel, unknown>;
}